CreateEvent.js working now
