+++ public/admin_home.html
@@ -9,6 +9,7 @@
     <title>Admin Homepage</title>
     <link rel="stylesheet" href="css/admin_home.css">
     <script type="module" src="js/admin.js" defer></script>
+    <script type="module" src="js/createEvent.js" defer></script>
 </head>
 <body>
     <h1>Welcome, Admin</h1>
