+++ server.js
@@ -13,7 +13,7 @@ console.log('Server is starting');
 
 
 
-//const serviceAccountPath = path.resolve('./serviceAccountKey.json');
+// const serviceAccountPath = path.resolve('./serviceAccountKey.json');
 
 
 // if (!fs.existsSync(serviceAccountPath)) {
